// src/app/api/model/user-preferences-dto-in.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserPreferencesDtoIn { 
    /**
     * Email notification frequency. Parsed as EmailFrequency enum on the server.
     */
    communicationFrequency?: UserPreferencesDtoInCommunicationFrequencyEnum;
    darkModeEnabled?: boolean;
    gdprMarketingConsent?: boolean;
    /**
     * Preferred language code
     */
    language?: UserPreferencesDtoInLanguageEnum;
    notificationEmailEnabled?: boolean;
    notificationEmailPartnershipEnabled?: boolean;
    notificationEmailSupportEnabled?: boolean;
    notificationPartnershipEnabled?: boolean;
    notificationPushEnabled?: boolean;
    notificationSmsEnabled?: boolean;
    notificationSupportEnabled?: boolean;
    notificationSystemEnabled?: boolean;
    sharePhoneForPayments?: boolean;
    timezone?: string;
    twoFactorAuthenticationEnabled?: boolean;
}
export enum UserPreferencesDtoInCommunicationFrequencyEnum {
    IMMEDIATE = 'IMMEDIATE',
    HOURLY_DIGEST = 'HOURLY_DIGEST',
    DAILY_DIGEST = 'DAILY_DIGEST',
    WEEKLY_DIGEST = 'WEEKLY_DIGEST'
};
export enum UserPreferencesDtoInLanguageEnum {
    PL = 'pl',
    EN = 'en'
};